Using a dedicated datastructure for side effect ordering.
We were doing fishy things in the Term_typing file, where side-effects were
not considered in the right uniquization order because of the uniq_seff_rev
function. It probably did not matter after a9b76df because effects were
(mostly) uniquize upfront, but this is not clear because of the use of the
transparente API in the module.

Now everything has to go through the opaque API, so that a proper dependence
order is ensured.