[stm] Improve error messages on add/parse.
As suggested by @psteckler in #524 , we give more explicit information
about what is wrong.

We also provide some debug information for the possible dangerous case
of having the tip go out of sync with the real installed state (which
will make parsing fail if there was some changes to the parser).

We also fix a couple of typos noticed by Paul, thanks Paul.