[stm] Make newtip mandatory on Stm.add
We force STM clients to pick the new state in documents. It is their
responsibility for the states to be unique. This has significant
advantages for IDE clients, as they don't have to wait now.

It also cleanups a bit of code and avoids some weird errors (cf #527,
Bugzilla#5101).

Note that the stm still generates internal states in some cases,
however, we clean up that part a by ensuring that all the private
states are negative, so IDEs can detect when feedback for an internal
state is generated.

The XML part of the protocol is not yet implemented, will be rolled
with the general version update there.