Fix substitution of abstracted lemmas.
Instead of browsing the term as many times as there are abstracted constants,
we replace the constants in one pass. We have to be a bit careful to replace
the right variables though, in case there are chained abstracts. This is much
faster.

This solves the second part of bug #5382: Huge case analysis fails in coq8.5.x.