[ide] Correctly place warning tags.
8e07227c5853de78eaed4577eefe908fb84507c0 introduced an
incorrect duplicate of `position_error_tag_at_sentence`, which
sets the end of the underlining position starting at the end of the
sentence, whereas the location in the feedback refers to the
beginning, thus it highlights more text than it should.

This was missed in 8.6 as it seems that the code was not called.
We undo the duplication and fix the bug.