[proof] Rework `proof_mode` handling.
This patch reworks "proof_mode" handling, moving the state itself to
the parsing summary, and giving the decision making to the STM.

When parsing, the STM will try to classify statically the start of a
proof. If the command can be statically classified, the STM will set
the parsing proof mode itself and allow us to continue parsing.

If the STM cannot classify the proof mode, it will indeed run the
command, and `Lemmas.start_proof/save_proof` will be the ones
doing the push/pop of proof mode states.