[ide] Rely less on `Stateid.dummy`
In particular, we set queries from the menu to the tip of the
document, and process feedback coming with a `dummy` id.

There are still more places to tweak, but this should be good for now.

We also display a few more query messages, in particular the feedbacks
produced by query that carry a dummy state id.

This hack of reporting with from the STM should be solved once we
update the protocol.