Answer to db28e827d: I did test the commit on test-suite but for some reason I overlooked the result, maybe a missing make clean.
As for testing on the travis test architecture, I could have done
it. This is a question of compromise. I was so certain that it would
work that I did not test. And anyway, the travis test is not absolute
either.

In any case, I'm very sorry about the confusion it introduced.

~~~~ BY THE WAY, NO NEED TO SHOUT ~~~~
~~~~ IT IS A BIG MISCONCEPTION ABOUT HUMAN BEINGS ~~~~
~~~~ TO BELIEVE THAT THEY DO MISTAKES INTENTIONALLY ~~~~

    Thank you!