Factorizing interp_gen through a function interpreting glob_constr.
The new function is interp_glob_closure which is basically a renaming
and generalization of interp_uconstr.

Note a change of semantics that I could however not observe in
practice.

Formerly, interp_uconstr discarded ltac variables bound to names for
interning, but interp_constr did not. Now, both discard them.

We also export the new interp_glob_closure.