Reorganize comment documentation of ChoiceFacts.v
Shortnames and natural language descriptions of principles are moved
next to each principle.
The table of contents is moved to after the principle definitions.
Extra definitions are moved to the definition section (eg DependentFunctionalChoice)
Compatibility notations have been moved to the end of the file.

Details:

The following used to be announced but were neither defined or used,
and have been removed:
- OAC!
- Ext_pred          = extensionality of predicates
- Ext_fun_prop_repr = choice of a representative among extensional functions to Prop

GuardedFunctionalRelReification was announced with shortname GAC! but
shortname GFR_fun was used next to it. Only the former has been retained.

Shortnames and descriptions have been invented for
InhabitedForallCommute DependentFunctionalRelReification
ExtensionalPropositionRepresentative ExtensionalFunctionRepresentative

Some modification of headlines