[ide] Disable `print_ast` call.
So far this part of the system has shown little utility other than
having developers put time to fix it every time they change something
in the system.

I have never seen this functionality used in the wild, and a large
part of the vernac was marked TODO. Given that we have automatic
methods to provide this functionality these days (PPX), we remove
Texmacspp.