Refine test for unresolved evars to be less sensitive to unification order
evar-evar unifications and restrictions can instantiate pre-existing
evars by new evars but these are not considered fresh
anymore. Similarly, fresh evars can be instantiated by pre-existing
evars and these are not considered fresh either for tactics not
supporting evars.

Add regression tests to the test-suite and a note explaining this in the
reference manual.