[api] Put some order in API files.
We start to sort out the dependency mess of API by providing a logical
declaration order in `API.{ml,mli}` that is related to the linking
order.

Due to the correct declaration order, we can remove now the spurious
module `API.Prelude` as well as some spurious declarations present
there.

Unfortunately we don't solve the main problems with `ml` files yet,
however this should be orthogonal to that problem. It is becoming
clear that OCaml doesn't really support well having "`mli`-only"
files.

We also had to move a bunch of modules (`Impargs`, `Declare`,
`Ind_tables`, `Miscprint`) to their proper place as they were linked
differently, for instance as they now depend on `EConstr` or the "new"
proof engine depends on them.