[api] Generate instance for `mli`-only modules for API export.
This is an alternative to #797 that should solve the current API
module problem.

IMO the ideal thing would be to get rid of `intf` and turn the `mli`
files into "proper" modules that can be aliases, but that'd be clearly
8.8 material.

We should still be more strict on the intended linking order when
declaring the modules in API, but this will work for now for Coq, but
may need some tweak in the order for external plugins.