Registering user-given names at goal initialization and in "assert".
This probaby works in usual cases but it is not absolute. E.g., if one
"assert" a lemma build from pieces obtaine by "match goal", we may use
names which would be considered user-given while they are not.  Also,
the test is done in internal_cut but other tactics than "assert" are
using internal_cut (and not necessarily with a "by" clause).