Do not add original constraints to the environment in 'with Definition' check.
This was useless, because immediate constraints are assumed to already be in
the current environment, while deferred constraints are useless for the
conversion check of the definition types, as they only appear in the opaque
body.

This also clarifies a bit what is going on in the typing of module constraints
w.r.t. global universes.