Removing dead code in Subtyping.
This code was a sketch of what to do when we properly implement module-level
handling of instanciation of definitions by inductive types. It was completely
dead code, called after an error, and somewhat incorrect. Instead of letting
it bitrot, we remove it.

(cherry picked from commit 78f536c7fa1af8a61c3dbc5eafae74ad436958ef)