[travis] Display info on tested commit for PR builds.
This is made necessary by the fact that the displayed commit in the header
of the build can be outdated if the build was restarted or took a long time
to start. Sometimes, it is so old that the link to GitHub is a 404.

(cherry picked from commit 8d895c065b9a3dc9cc54b0cbc0f6f8daf275ece4)