Getting rid of abstraction breaking code in tclABSTRACT.
This is probably the hardest case of them all, because tclABSTRACT fundamentally
relies on the names of universes from the constant instance being the same as
the one in the current goal. Adding to that the fact that the kernel is doing
strange things when provided with a polymorphic definition with body universe
constraints, it turns out to be a hellish nightmare to handle properly.

At some point we need to clarifiy this in the kernel as well, although we
leave it for some other patch.