Fixing what was presumably a typo in the naming conventions file.
Indeed, "forall x, op x x = x" in not in the list, while this is one
of the two standard meanings of idempotence. So, knowing that x, y,
... and not n are used elsewhere for variables names, and elt for
constants. Moreover, it is probable that before using consistently x,
y and z, I had also used m and n, sometimes. So, a convergent
probability that it is (just) a typo.