[api] Put some order in API.mli
We try to sort out the dependency graph of API by following a logical
declaration order in `API.{ml,mli}` related to the actual dependency
order of Coq modules.

Things are a bit tricky here as Coq itself relies on the fact that
OCaml treats module interface and implementation separately
dependency-wise; however, when resorting module alias the design seems
to become more coupled.

Currently, API exposes both "namespaces", asserting a large number of
type equality between them, however the `API` namespace is not
self-contained.

In particular, try to solve problems such as `Summary.frozen` being
used in `API.mli` but not declared by the `Summary` module exported by
API, etc... In general we try to follow the invariant that a type used
in `API` has been declared before.

Keep in mind that OCaml upstream has warned that it maybe tricky to
alias objects in this way. In particular, after API the old `mli` only
files have become full compilation units so we may want to be more
careful here.

The more "correct" declaration order allows us to remove the
`API.Prelude` module, as well as some other declarations that I
consider as spurious. However we still maintain the large number of
type aliases.

We follow the linking order except for the files in `intf`, which are
conceptually wrongly placed in the linking hierarchy but this doesn't
matter as the files don't contain any implementation.

The new location in `API.mli` do indicate where do they really sit in
the dependency chain.

Other change is that a couple of `.mli` only files have been turned
into `.ml`-only files to ease the module aliasing in OCaml 4.02.3

We still keep some `ml`-only files, it seems that OCaml doesn't
support well having "`mli`-only" files + module aliases, so what to do
with this question and API is open.

The minor changes to `open` are due to scope warnings, which were done
by the improper aliasing of `Term` to `Constr`.