[api] Split API in logical files.
We split the API in their "logical" components to ease the handling of
the API.mli file, and hopefully facilitate the life of each subsystem
maintainer.

This commit is not yet ready, but it already makes the coupling
between components more explicit, and indeed from small experiments I
can attests that the coupling could be greatly improved with minor
changes to the codebase.